requirements.txt
absl-py==0.10.0
argcomplete==1.12.1
astor==0.8.1
astroid==2.4.2
autopep8==1.5.4
boto3==1.16.4
botocore==1.19.4
certifi==2020.6.20
cffi==1.14.3
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
Cython==0.29.21
durationpy==0.5
Flask==1.1.2
Flask-API==2.0
future==0.18.2
gast==0.2.2
gevent==20.9.0
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.33.1
h5py==2.10.0
hjson==3.0.2
idna==2.10
importlib-metadata==2.0.0
install==1.3.4
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
Keras==2.3.0
# Keras-Applications==1.0.8
# Keras-Preprocessing==1.1.2
keras-resnet==0.1.0
keras-retinanet==0.5.1
lazy-object-proxy==1.4.3
libmagic==1.0
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.2
opencv-python==4.4.0.44
opt-einsum==3.3.0
Pillow==8.0.1
pip-tools==5.3.1
placebo==0.9.0
progressbar2==3.53.1
protobuf==3.13.0
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
python-magic==0.4.18
python-slugify==4.0.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
scipy==1.5.3
six==1.15.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.1
tqdm==4.51.0
troposphere==2.6.3
typed-ast==1.4.1
urllib3==1.25.11
Werkzeug==0.16.1
wrapt==1.12.1
wsgi-request-logger==0.4.6
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2
